[VPlan] Move widening decisions to VPlanWideningDecisions.cpp (NFC)
Following up to, move transformations making decisions about widening to
VPlanWideningDecisions.cpp.

This moves
tryToConvertVPInstructionsToVPRecipes, convertToAbstractRecipes,
createPartialReductions, makeMemOpWideningDecisions, makeScalarizationDecisions,
makeCallWideningDecisions and convertToStridedAccesses together with their
static helpers.
