+++ llvm/lib/Transforms/Vectorize/CMakeLists.txt
@@ -41,6 +41,7 @@ add_llvm_component_library(LLVMVectorize
   VPlanUnroll.cpp
   VPlanVerifier.cpp
   VPlanUtils.cpp
+  VPlanWideningDecisions.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/Transforms
